#96 Use waiters from inelegant.net; remove local waiters.
We moved the "waiter functions" (functions that block until a port is
responding/stopped to respond) to the Inelegant package. Let's use them.

Note that the server code is way behind the ideal now: it should be able to be
used without the Inelegant package. We have to improve it soon.