Reparing "eager prefix replacement"... and finding anoter problem.
We resolved the problem from the last commit by replacing only the first
occurrence of the prefix. Yet we still have a issue. We assume that there will
be a slash at the beginning of the second part, but that does not happen if
the prefix contains a slash itself::

    >>> split_path('/a/b/c', {'/a/'})
    ('/a/', 'b/c')

Of course, this assumption fails all the time if the default prefix '/' is
used::

    >>> split_path('/a/b/c', {})
    ('/', 'a/b/c')

Did you know that?! I didn't!